refactor: extract buildCorrectionContext into dedicated module
Moves the AI correction-context builder out of useTransactionHandlers into
src/services/ai/correction-context.ts so it can be reused by the AI
categorization preview dev tool without going through the hook.
